# docs/openapi.yaml
openapi: 3.0.3
info:
  title: India In-Time API
  version: "3.6.0"
  description: |
    Production API contract for India In-Time. Request and response schemas are
    intentionally explicit so frontend, mobile, QA, and third-party clients can
    validate payloads without reverse-engineering route implementations.
servers:
  - url: /api
    description: Deployment-relative API root
...
paths:
...
  /places:
    get:
      tags: [Places]
      summary: Discover places for a city
      parameters:
        - { name: city, in: query, required: false, schema: { type: string, maxLength: 100 } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200':
          description: Place list
          content:
            application/json:
              schema:
                type: object
                properties:
                  city: { type: string }
                  places:
                    type: array
                    items: { $ref: '#/components/schemas/Place' }
                required: [places]
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /favorites:
    get:
      tags: [Favorites]
      summary: List current user's active favorites
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorites:
                    type: array
                    items: { $ref: '#/components/schemas/Favorite' }
                required: [favorites]
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Favorites]
      summary: Add or restore a favorite
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FavoriteCreate' }
      responses:
        '201':
          description: Favorite created or restored
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Favorite' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /favorites/{id}:
    parameters:
      - { $ref: '#/components/parameters/FavoriteId' }
    delete:
      tags: [Favorites]
      summary: Soft-delete a favorite
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /feedback/place:
    post:
      tags: [Feedback]
      summary: Submit place feedback
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlaceFeedbackCreate' }
      responses:
        '201':
          description: Feedback recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeedbackCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      tags: [Feedback]
      summary: Admin place feedback summary
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Feedback summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeedbackSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /analytics/gemini:
    get:
      tags: [Analytics]
      summary: Gemini usage and estimated cost
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: hours, in: query, schema: { type: integer, minimum: 1, maximum: 720, default: 24 } }
      responses:
        '200':
          description: Usage summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GeminiAnalytics' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /analytics/ml/crowd:
    get:
      tags: [Analytics]
      summary: Crowd ML model information
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Model information
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CrowdModelInfo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /analytics/ml/crowd/train:
    post:
      tags: [Analytics]
      summary: Train crowd model from authenticated feedback
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                maxRows: { type: integer, minimum: 1, maximum: 100000, default: 10000 }
      responses:
        '200':
          description: Training result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrainingResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: Training already running }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  parameters:
...
    FavoriteId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        requestId: { type: string }
...
    Coordinates:
      type: array
      minItems: 2
      maxItems: 2
      items: { type: number }
      example: [17.6868, 83.2185]
...
    Place:
      type: object
      required: [id, name, category, coords]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        coords: { $ref: '#/components/schemas/Coordinates' }
        visitTimeMinutes: { type: integer, minimum: 0 }
        openTime: { type: string, pattern: '^\\d{2}:\\d{2}$' }
        closeTime: { type: string, pattern: '^\\d{2}:\\d{2}$' }
        rating: { type: number, minimum: 0, maximum: 5 }
        importance: { type: string }
...
    FavoriteCreate:
      type: object
      required: [placeName, city]
      properties:
        placeName: { type: string, minLength: 1, maxLength: 200 }
        city: { type: string, minLength: 1, maxLength: 100 }
        category: { type: string, maxLength: 50 }
        coords: { $ref: '#/components/schemas/Coordinates' }
        notes: { type: string, maxLength: 2000 }
    Favorite:
      allOf:
        - $ref: '#/components/schemas/FavoriteCreate'
        - type: object
          required: [id]
          properties:
            id: { type: integer, format: int64 }
            createdAt: { type: string, format: date-time }
...
    PlaceFeedbackCreate:
      type: object
      required: [placeName, rating]
      properties:
        placeName: { type: string, minLength: 1, maxLength: 200 }
        city: { type: string, maxLength: 100 }
        rating: { type: integer, minimum: 1, maximum: 5 }
        tags:
          type: array
          maxItems: 10
          items: { type: string, maxLength: 50 }
        comment: { type: string, maxLength: 2000 }
...
    FeedbackCreated:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        id: { type: string }
    FeedbackSummary:
      type: object
      properties:
        count: { type: integer, minimum: 0 }
        averageRating: { type: number, minimum: 0, maximum: 5 }
        items:
          type: array
          items: { type: object, additionalProperties: true }
    GeminiAnalytics:
      type: object
      properties:
        hours: { type: integer, minimum: 1 }
        requests: { type: integer, minimum: 0 }
        failures: { type: integer, minimum: 0 }
        estimatedCostUsd: { type: number, minimum: 0 }
        cachedResponses: { type: integer, minimum: 0 }
    CrowdModelInfo:
      type: object
      properties:
        version: { type: string }
        trainedAt: { type: string, format: date-time, nullable: true }
        sampleCount: { type: integer, minimum: 0 }
        features: { type: array, items: { type: string } }
    TrainingResult:
      type: object
      properties:
        success: { type: boolean }
        modelVersion: { type: string }
        samplesUsed: { type: integer, minimum: 0 }
        startedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }